requirements.txt
Pillow==9.5.0
Pillow==11.3.0
Requests==2.32.4
